telegram_gateway/requirements.txt
telethon
requests
python-dotenv


